requirements.txt
streamlit
Pillow
numpy
torch
torchvision
dominate
visdom
packaging
GPUtil
